backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = ""
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.115.6",
    "pydantic-settings>=2.7.1",
    "pyjwt>=2.10.1",
    "passlib[bcrypt]<2.0.0,>=1.7.4",
    "sqlmodel>=0.0.22",
    "uvicorn>=0.34.0",
    "pydantic[email]>=2.10.4",
    "pymysql>=1.1.1",
    "python-multipart>=0.0.20",
    "tenacity>=9.0.0",
    "alembic>=1.14.0",
    "cryptography>=44.0.0",
    "bcrypt==4.0.1",
    "jinja2>=3.1.5",
    "emails>=0.6",
    "minio>=7.2.14",
    "ultralytics>=8.3.59",
]
